# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface QuizInput {
  /** @minLength 1 */
  race: string;
}

export interface QuizQuestion {
  q: string;
  opts: string[];
  ans: number;
  fact: string;
}

export interface QuizResult {
  race: string;
  questions: QuizQuestion[];
}

export interface ApiError {
  error: string;
}