# lib/api-zod/src/generated/types/quizResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from './quizQuestion';

export interface QuizResult {
  race: string;
  questions: QuizQuestion[];
}